+++ server/api/r2/upload.ts
@@ -1,5 +1,5 @@
 import { readMultipartFormData, createError } from "h3";
-import { smmsUpload } from "../../../app/utils/api/smms";
+import { r2Upload } from "../../../app/utils/api/r2";
 
 export default defineEventHandler(async (event) => {
   if (event.node.req.method?.toUpperCase() !== "POST") {
@@ -10,11 +10,23 @@ export default defineEventHandler(async (event) => {
   }
   try {
     const form = await readMultipartFormData(event);
-    const token = form!.find(i => i.name === "token")?.data.toString();
+    const secretAccessKey = form!.find(i => i.name === "secretAccessKey")?.data.toString();
     const tinyPngToken = form!.find(i => i.name === "tinyPngToken")?.data.toString();
     const file = form!.find(i => i.name === "file");
 
-    return await smmsUpload({ token, tinyPngToken, file });
+    if (!file || !file.data) {
+      throw new Error("No file provided");
+    }
+
+    return await r2Upload({
+      secretAccessKey,
+      tinyPngToken,
+      file: {
+        data: file.data,
+        filename: file.filename,
+        type: file.type
+      }
+    });
   } catch (e: any) {
     return createError({
       statusCode: 503,